+++ CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - OTP 18 support
 - OTP 19.0, 19.1 and 19.2 support
 - ability of consulting database metadata, source and version through `:get_info`
+- ability of subscribing database loader events